hrms/requirements.txt
Django==5.2.11
djangorestframework==3.15.1
django-cors-headers==4.9.0
gunicorn==22.0.0
